devices/scanner/mindeo_p80/requirements.txt
aiofiles==24.1.0
aioserial==1.3.1
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.7.14
evdev==1.9.2
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyhumps==3.8.0
pynput==1.8.1
pyserial==3.5
python-xlib==0.33
redis==6.2.0
six==1.17.0
sniffio==1.3.1
typing-extensions==4.14.1
